+++ src/api/v1/index.ts
@@ -1,12 +1,19 @@
 import { zValidator } from "@hono/zod-validator";
-import { and, desc, eq, lt } from "drizzle-orm";
+import { and, desc, eq, gte, inArray, lt, lte } from "drizzle-orm";
 import { Hono } from "hono";
 import { z } from "zod";
 import { db } from "../../db";
+import { serializeAccount } from "../../entities/account";
 import { getPostRelations, serializePost } from "../../entities/status";
 import { serializeTag } from "../../entities/tag";
 import { type Variables, scopeRequired, tokenRequired } from "../../oauth";
-import { bookmarks, likes } from "../../schema";
+import {
+  accounts as accountsTable,
+  bookmarks,
+  likes,
+  mutes,
+} from "../../schema";
+import { uuid } from "../../uuid";
 import accounts from "./accounts";
 import apps from "./apps";
 import featured_tags from "./featured_tags";
@@ -197,6 +204,61 @@ app.get(
     return c.json(
       owner.followedTags.map((tag) => serializeTag(tag, owner, c.req.url)),
     );
+  },
+);
+
+app.get(
+  "/mutes",
+  tokenRequired,
+  scopeRequired(["read:mutes"]),
+  zValidator(
+    "query",
+    z.object({
+      max_id: uuid.optional(),
+      since_id: uuid.optional(),
+      limit: z
+        .string()
+        .default("40")
+        .transform((v) => {
+          const parsed = Number.parseInt(v);
+          return Math.min(parsed, 80);
+        }),
+    }),
+  ),
+  async (c) => {
+    const owner = c.get("token").accountOwner;
+    if (owner == null) {
+      return c.json(
+        { error: "This method requires an authenticated user" },
+        422,
+      );
+    }
+
+    const muteList = await db.query.mutes.findMany({
+      where: eq(mutes.accountId, owner.id),
+    });
+
+    if (muteList.length < 1) return c.json([]);
+
+    const query = c.req.valid("query");
+
+    const mutedAccounts = await db.query.accounts.findMany({
+      where: and(
+        inArray(
+          accountsTable.id,
+          muteList.map((m) => m.mutedAccountId),
+        ),
+        query.max_id == null ? undefined : lte(accountsTable.id, query.max_id),
+        query.since_id == null
+          ? undefined
+          : gte(accountsTable.id, query.since_id),
+      ),
+      with: { owner: true, successor: true },
+      orderBy: [desc(accountsTable.id)],
+      limit: query.limit ?? 40,
+    });
+
+    return c.json(mutedAccounts.map((a) => serializeAccount(a, c.req.url)));
   },
 );