+++ CHANGES.md
@@ -15,11 +15,17 @@ To be released.
      -  When a `Announce(Note)` or `Undo(Announce(Note))` activity is received,
         now it is forwarded to the followers as well if the activity is signed.
 
+ -  Added [`GET /api/v1/mutes`] API to Mastodon comapatiblity layer.  This API
+    returns a list of accounts that are muted by the authenticated user.
+    [[#103]]
+
  -  On profile page, images are no more captioned using `<figcaption>` but
     use only `alt` attribute for accessibility.  [[#99], [#100] by Okuto Oyama]
 
 [#99]: https://github.com/fedify-dev/hollo/issues/99
 [#100]: https://github.com/fedify-dev/hollo/pull/100
+[#103]: https://github.com/fedify-dev/hollo/issues/103
+[`GET /api/v1/mutes`]: https://docs.joinmastodon.org/methods/mutes/#get
 
 
 Version 0.4.5